#9 Change the delineation strategy from fixed length to encoded length to be able to easily tweak the size of the messages without a need to modify the echo application

+++ asynctcpacceptance/src/main/java/dev/squaremile/asynctcpacceptance/AdHocProtocol.java
@@ -2,6 +2,6 @@ package dev.squaremile.asynctcpacceptance;
 
 public class AdHocProtocol
 {
-    public static final long NO_OPTIONS = 0L;
-    public static final long PLEASE_RESPOND_FLAG = 1L;
+    public static final int NO_OPTIONS = 0;
+    public static final int PLEASE_RESPOND_FLAG = 1;
 }
